Ignore whole .idea directory
Upon importing a freshly cloned repo Android Studio creates a bunch of config files in .idea directory. For me there was one because I clicked on the suggestion to add a VCS root. Another because I fixed the run configuration (by default it tries to launch an activity, but there is none here, so the default run config was not functional). There's a bunch more of them that just spawned automatically without any action on my part.

I am assuming there is no intention to sync this configuration between clones, seeing as it isn't provided in the repository. I agree with this choice and with this change suggest to simply ignore the whole .idea directory and let each Android Studio instance generate it's own config files.